refactor password dialogs
