+++ packages/atlas/src/components/_auth/PasswordCriterias/PasswordCriterias.styles.ts
@@ -3,11 +3,6 @@ import styled from '@emotion/styled'
 import { Text } from '@/components/Text'
 import { sizes } from '@/styles'
 
-export const StyledSignUpForm = styled.form`
-  display: grid;
-  gap: ${sizes(6)};
-`
-
 export const PasswordRequirementsWrapper = styled.section`
   display: grid;
   gap: ${sizes(2)};
