+++ packages/atlas/src/components/_auth/PasswordCriterias/index.ts
@@ -1,0 +1,1 @@
+export * from './PasswordCriterias'
