+++ packages/atlas/src/views/viewer/MembershipSettingsView/ChangePasswordDialog.tsx
@@ -1,10 +1,15 @@
 import styled from '@emotion/styled'
+import { zodResolver } from '@hookform/resolvers/zod'
 import { u8aToHex } from '@polkadot/util'
 import axios from 'axios'
 import { FC, useCallback, useRef, useState } from 'react'
+import { FormProvider, useForm } from 'react-hook-form'
+import { z } from 'zod'
 
-import { PasswordForm } from '@/components/_auth/PasswordForm/PasswordForm'
-import { DialogButtonProps } from '@/components/_overlays/Dialog'
+import { SvgActionHide, SvgActionShow } from '@/assets/icons'
+import { PasswordCriterias } from '@/components/_auth/PasswordCriterias'
+import { FormField } from '@/components/_inputs/FormField'
+import { Input, InputProps } from '@/components/_inputs/Input'
 import { DialogModal } from '@/components/_overlays/DialogModal'
 import { atlasConfig } from '@/config'
 import { ORION_AUTH_URL } from '@/config/env'
@@ -12,7 +17,7 @@ import { keyring } from '@/joystream-lib/lib'
 import { decodeSessionEncodedSeedToMnemonic, prepareEncryptionArtifacts } from '@/providers/auth/auth.helpers'
 import { useAuth } from '@/providers/auth/auth.hooks'
 import { useSnackbar } from '@/providers/snackbars'
-import { media } from '@/styles'
+import { media, sizes } from '@/styles'
 import { SentryLogger } from '@/utils/logs'
 
 type ChangePasswordArgs = {
@@ -63,21 +68,57 @@ export const changePassword = async ({
   }
 }
 
+const commonPasswordValidation = z
+  .string()
+  .regex(/^(?=.*[0-9])(?=.*[A-Z])(?=.*[!@#$%^&*()_+]).*$/, { message: 'Password has to meet requirements.' })
+  .min(9, { message: 'Password has to meet requirements.' })
+
+const zodSchema = z
+  .object({
+    password: commonPasswordValidation,
+    confirmPassword: commonPasswordValidation,
+  })
+  .refine(
+    (data) => {
+      return data.password === data.confirmPassword
+    },
+    {
+      path: ['confirmPassword'],
+      message: 'Password address has to match.',
+    }
+  )
+
 type ChangePasswordDialogProps = {
   onClose: () => void
   show: boolean
 }
 
+type PasswordStepForm = {
+  password: string
+  confirmPassword: string
+}
+
+type PasswordInputNames = keyof PasswordStepForm
+
 export const ChangePasswordDialog: FC<ChangePasswordDialogProps> = ({ onClose, show }) => {
-  const [primaryButtonProps, setPrimaryButtonProps] = useState<DialogButtonProps>()
   const { displaySnackbar } = useSnackbar()
+
+  const form = useForm<PasswordStepForm>({
+    shouldFocusError: true,
+    resolver: zodResolver(zodSchema),
+  })
+
+  const [isFieldVisible, setIsFieldVisible] = useState<Record<PasswordInputNames, boolean>>({
+    password: false,
+    confirmPassword: false,
+  })
 
   const [isSubmitting, setIsSubmiting] = useState(false)
   const dialogContentRef = useRef<HTMLDivElement>(null)
   const { currentUser, encodedSeed } = useAuth()
 
-  const handleChangePassword = useCallback(
-    async (password: string) => {
+  const handleChangePassword = useCallback(() => {
+    form.handleSubmit(async (data) => {
       if (!currentUser || !encodedSeed) {
         return
       }
@@ -86,7 +127,7 @@ export const ChangePasswordDialog: FC<ChangePasswordDialogProps> = ({ onClose, s
         await changePassword({
           email: currentUser.email,
           encodedSeed,
-          newPassword: password,
+          newPassword: data.password,
           gatewayAccountId: currentUser.id,
           joystreamAccountId: currentUser.joystreamAccount,
         })
@@ -104,8 +145,28 @@ export const ChangePasswordDialog: FC<ChangePasswordDialogProps> = ({ onClose, s
       } finally {
         setIsSubmiting(false)
       }
+    })()
+  }, [currentUser, displaySnackbar, encodedSeed, form, onClose])
+
+  const hasDoneInitialScroll = useRef(false)
+
+  const handleTogglePassword = (name: PasswordInputNames) => {
+    setIsFieldVisible((fields) => ({ ...fields, [name]: !fields[name] }))
+  }
+  const getInputProps = useCallback(
+    (name: PasswordInputNames): InputProps => {
+      return {
+        ...form.register(name),
+        type: isFieldVisible[name] ? 'text' : 'password',
+        actionButton: {
+          children: isFieldVisible[name] ? 'Hide' : 'Show',
+          dontFocusOnClick: true,
+          icon: isFieldVisible[name] ? <SvgActionHide /> : <SvgActionShow />,
+          onClick: () => handleTogglePassword(name),
+        },
+      }
     },
-    [currentUser, displaySnackbar, encodedSeed, onClose]
+    [form, isFieldVisible]
   )
 
   return (
@@ -115,21 +176,35 @@ export const ChangePasswordDialog: FC<ChangePasswordDialogProps> = ({ onClose, s
       title="Change password"
       contentRef={dialogContentRef}
       primaryButton={{
-        ...primaryButtonProps,
         text: isSubmitting ? 'Waiting...' : 'Change password',
         disabled: isSubmitting,
+        onClick: handleChangePassword,
       }}
       secondaryButton={{
         text: 'Cancel',
         onClick: onClose,
       }}
     >
-      <PasswordForm
-        type="change"
-        onPasswordSubmit={handleChangePassword}
-        setPrimaryButtonProps={setPrimaryButtonProps}
-        dialogContentRef={dialogContentRef}
-      />
+      <Wrapper>
+        <FormField label="New password" error={form.formState.errors.password?.message}>
+          <Input
+            placeholder="New password"
+            {...getInputProps('password')}
+            autoComplete="off"
+            onClick={() => {
+              if (hasDoneInitialScroll.current || !dialogContentRef?.current) return
+              hasDoneInitialScroll.current = true
+              dialogContentRef.current.scrollTo({ top: dialogContentRef.current.scrollHeight, behavior: 'smooth' })
+            }}
+          />
+        </FormField>
+        <FormField label="Repeat Password" error={form.formState.errors.confirmPassword?.message}>
+          <Input placeholder="Repeat password" {...getInputProps('confirmPassword')} autoComplete="off" />
+        </FormField>
+        <FormProvider {...form}>
+          <PasswordCriterias />
+        </FormProvider>
+      </Wrapper>
     </StyledDialogModal>
   )
 }
@@ -139,4 +214,9 @@ const StyledDialogModal = styled(DialogModal)`
   ${media.sm} {
     max-height: 576px;
   }
+`
+
+const Wrapper = styled.div`
+  display: grid;
+  gap: ${sizes(6)};
 `
